# packages/purebackend/docs/swagger.yaml
definitions:
...
  models.LogRequest:
    properties:
      data:
        type: string
      key:
        type: string
    type: object
...
info:
  contact:
    email: contact@pureml.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}/log:
    get:
      consumes:
      - '*/*'
      description: Get Log data for dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Log data for dataset
      tags:
      - Dataset
    post:
      consumes:
      - '*/*'
      description: Log data for dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      - description: Data to log
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Log data for dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}/log/{key}:
    get:
      consumes:
      - '*/*'
      description: Get Log data for dataset with specific key
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      - description: Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Log data for dataset with specific key
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}/logfile:
    post:
      consumes:
      - multipart/form-data
      description: Log file data for dataset
      parameters:
      - description: Dataset files (multiple supported)
        in: formData
        items:
          type: file
        name: file
        required: true
        type: array
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      - in: formData
        name: storage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Log file data for dataset
      tags:
      - Dataset
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Header for logged in user format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"